src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
backcall=0.1.0=py37_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37he30daa8_1
chardet=3.0.4=py37_1003
colorama=0.4.3=pypi_0
cryptography=2.9.2=py37h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=pypi_0
cymem=1.31.2=py37h6bb024c_0
cytoolz=0.9.0.1=py37h14c3975_1
dbus=1.13.14=hb2f20db_0
decorator=4.4.2=py_0
dill=0.2.9=py37_0
ekphrasis=0.5.1=pypi_0
en-core-web-sm=2.0.0=pypi_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
ftfy=5.7=pypi_0
glib=2.63.1=h3eb4bd4_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
idna=2.9=py_1
intel-openmp=2020.1=217
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
jedi=0.17.0=py37_0
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py37hfd86e86_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libopenblas=0.3.6=h5a2b251_2
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
matplotlib=3.2.1=pypi_0
matplotlib-base=3.1.3=py37hef1b27d_0
mkl=2020.1=217
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
msgpack-numpy=0.4.3.2=py37_0
msgpack-python=0.5.6=py37h6bb024c_1
murmurhash=0.28.0=py37hf484d3e_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
nltk=3.4.5=py37_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
openssl=1.1.1g=h7b6447c_0
pandas=1.0.3=py37h0573a6f_0
parso=0.7.0=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pip=20.0.2=py37_3
plac=0.9.6=py37_0
preshed=1.0.1=py37he6710b0_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
ptyprocess=0.6.0=py37_0
pycparser=2.20=py_0
pygments=2.6.1=py_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=py_0
pytorch=1.4.0=cuda101py37h02f0884_0
pytz=2020.1=py_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2018.07.11=py37h14c3975_0
requests=2.23.0=py37_0
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=py37h0b6359f_0
seaborn=0.10.1=py_0
setuptools=46.2.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.14.0=py37_0
spacy=2.0.12=py37h962f231_0
sqlite=3.31.1=h62c20be_1
termcolor=1.1.0=py37_1
thinc=6.10.3=py37h962f231_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
tornado=6.0.4=py37h7b6447c_1
tqdm=4.46.0=py_0
traitlets=4.3.3=py37_0
ujson=2.0.3=py37he6710b0_0
urllib3=1.25.8=py37_0
wcwidth=0.1.9=py_0
wheel=0.34.2=py37_0
wrapt=1.10.11=py37h14c3975_2
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
